# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for cloud-init clients using the OpenCHAMI cloud-init server",
        "title": "OpenCHAMI Cloud-Init Server API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/OpenCHAMI/.github/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/groups": {
            "get": {
                "description": "Get meta-data and cloud-init config for all groups known to\ncloud-init.  Note that group membership is managed outside of\nthe cloud-init service, normally in SMD.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "groups"
                ],
                "summary": "Get groups known by cloud-init",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/cistore.ClusterDefaults"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Add a new group to cloud-init corresponding to an SMD group.\nGroup-wide meta-data and/or a cloud-init configuration (in\neither plain or base64 encoding) can be specified.\n\nIf successful, a 201 Created status is returned and the\n`Location` header is set to the new group's groups endpoint,\n`/groups/{name}`.\n\nIf request parsing fails, a 422 Unprocessable Entity status is\nreturned. If adding group data to the data store fails, a 409\nConflict status is returned.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "groups"
                ],
                "summary": "Add a new group",
                "parameters": [
                    {
                        "description": "Group data",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cistore.GroupData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "/groups/{id}"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/admin/groups/{id}": {
            "get": {
                "description": "Get meta-data and cloud-init config for a single group known to\ncloud-init.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "groups"
                ],
                "summary": "Get data for single group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cistore.GroupData"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "Delete a group with its meta-data and cloud-init config.",
                "tags": [
                    "admin",
                    "groups"
                ],
                "summary": "Delete a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/admin/groups/{name}": {
            "put": {
                "description": "Set meta-data or cloud-init configuration for a specific group,\noverwriting any previous values.\n\nIf successful, a 201 Created status is returned and the\n`Location` header is set to the new group's groups endpoint,\n`/groups/{group}`. This operation is idempotent and replaces\nany existing content.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "groups"
                ],
                "summary": "Set group-specific meta-data and/or cloud-init config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Group data",
                        "name": "group_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cistore.GroupData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "/groups/{name}"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "cistore.CloudConfigFile": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "IyMgdGVtcGxhdGU6IGppbmphCiNjbG91ZC1jb25maWcKbWVyZ2VfaG93OgotIG5hbWU6IGxpc3QKICBzZXR0aW5nczogW2FwcGVuZF0KLSBuYW1lOiBkaWN0CiAgc2V0dGluZ3M6IFtub19yZXBsYWNlLCByZWN1cnNlX2xpc3RdCnVzZXJzOgogIC0gbmFtZTogcm9vdAogICAgc3NoX2F1dGhvcml6ZWRfa2V5czoge3sgZHMubWV0YV9kYXRhLmluc3RhbmNlX2RhdGEudjEucHVibGljX2tleXMgfX0KZGlzYWJsZV9yb290OiBmYWxzZQo="
                },
                "encoding": {
                    "type": "string",
                    "enum": [
                        "base64",
                        "plain"
                    ]
                },
                "filename": {
                    "type": "string"
                }
            }
        },
        "cistore.ClusterDefaults": {
            "type": "object",
            "properties": {
                "availability-zone": {
                    "type": "string"
                },
                "base-url": {
                    "type": "string",
                    "example": "http://demo.openchami.cluster:8081/cloud-init"
                },
                "boot-subnet": {
                    "type": "string"
                },
                "cloud_provider": {
                    "type": "string"
                },
                "cluster-name": {
                    "type": "string",
                    "example": "demo"
                },
                "nid-length": {
                    "type": "integer",
                    "example": 3
                },
                "public-keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIMLtQNuzGcMDatF+YVMMkuxbX2c5v2OxWftBhEVfFb+U user1@demo-head",
                        "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIB4vVRvkzmGE5PyWX2fuzJEgEfET4PRLHXCnD1uFZ8ZL user2@demo-head"
                    ]
                },
                "region": {
                    "type": "string"
                },
                "short-name": {
                    "type": "string",
                    "example": "nid"
                },
                "wg-subnet": {
                    "type": "string"
                }
            }
        },
        "cistore.GroupData": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "The compute group"
                },
                "file": {
                    "$ref": "#/definitions/cistore.CloudConfigFile"
                },
                "meta-data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "type": "string",
                    "example": "compute"
                },
                "versions": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}